backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.26.4
pyjwt==2.8.0
bcrypt==4.1.2
